requirements.txt
flask
numpy
pandas
scikit-learn
tensorflow-cpu
requests
streamlit